Manage conflicts with the last overriding the data.
I would have rather return validation error in case of any conflicts but
Django 'get' objects at each request and when we get, we sync with cdms
so conflicts don't really happen.
The last one overrides the version but it might not be a problem as we should
keep a copy of any change.