Sync modified field using CDMS as source of truth
The old logic used two different values for local_obj.modified and
cmds_obj.modifiedOn and checked that the delta between the two
was small enough (constant in settings) for the objs to be considered
syncronised.

The new logic gets rid of the delta and updates the local_obj.modified
with the value from cdms every time an operation happens.
In this way, we consider the two objs in sync if
local_obj.modified == cdms_obj.modifiedOn.