Make sure .all() call is covered
When Clazz.objects.all() gets called:
- cdms objs that don't exist in local, get created locally
- local objs out-of-date get updated from cdms ones

In case of exceptions during cdms calls, the error is propagated.

As cdms only supports 50 window size results and it's not efficient
to iterate over all the list every time, the developer has to acknowledge that
the .all() call might not include all the objects.
This is unfortunately a limitation imposed by the tech constraints.

Clazz.objects.skip_cdms().all() will skip cdms calls and only hit the local db